Updated chart parser:
- made wrap operation slightly more generally applicable (it now
applies to np\s whenever s is not a main sentence) the previous
strategy made it impossible for an adverb to have scope over an a+inf
construction, leading to problems with coordinations of such structures)

- added <x> export_stored command to the chart parse gui to output the
current parser state.

- added new check_solution predicate with an Index argument; this
checks the termination condition of the interactive parser.

- the “duplicate solution proof” problem has been solved; previously,
the interactive chart parser output two proofs of a solution.