Added all_sentences/0 predicate to check_proofs.pl
The predicate all_sentences/0 outputs the yields of the conclusions of
all proofs in the corpus. This is useful for comparison of the
extracted treebank with the original (using the maxent_words script of
the TLGbank).

In addition, the files m2const and transform_proof have been updated to
include the two new files monde_a and monde_b