Updated with "hybrid" probabilities
crosses/4 information is now converted to a probability penalty, where
a crossing number of C is translated to a probability of 1/(C+1). (the
log of this value is taken and added, since we are using
log-probabilities and addition instead of probabilities and
multiplication). This is experimental and needs proper evaluation.