Updated TclTk location
This is a workaround to get the right TclTk version in MacOS. This is currently flagged as deprecated, though. An alternative solution needs to be found.

Add new binding for Supertag.tcl where clicking <Mouse-1> on a formula/probability will display both; this has been added because the <Enter> binding is (for some reason) unreliable when using an external screen or display.